% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pell-package.R
\docType{package}
\name{pell-package}
\alias{pell}
\alias{pell-package}
\title{pell: An R package containing data about historic Pell grant distribution in the US}
\description{
Historic Pell grant data as provided by the US Department of Education. This package contains data about how much pell grant was awarded by which institution in which year.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Curious-Joe/pell}
  \item Report bugs at \url{https://github.com/Curious-Joe/pell/issues}
}

}
\author{
\strong{Maintainer}: Arafath Hossain \email{a.h.fahad90@gmail.com}

}
\keyword{internal}
