% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pell_grant_data.R
\docType{data}
\name{pell}
\alias{pell}
\title{Pell Grant Data}
\format{
A data frame with six variables:
\describe{
\item{\code{STATE}}{name of the state of the respective institution}
\item{\code{AWARD}}{total dollar value of awarded pell grant by the respective institution}
\item{\code{RECIPIENT}}{total number of pell grant recipients at that institution}
\item{\code{NAME}}{name of the college/university}
\item{\code{SESSION}}{Academic session when the grant was awarded}
\item{\code{YEAR}}{Fiscal year of when the grant was awarded}
}

For further details about the federal Pell grant program, see \url{https://www2.ed.gov/programs/fpg/index.html}
}
\usage{
pell
}
\description{
Total dollar amount awarded as Pell grant by different colleges over the years.
}
\keyword{datasets}
